ver 0.0.12 初始框架搭建 用户界面搭建,完成个人收藏页面功能

+++ src/com/cykj/dao/ICourseDao.java
@@ -12,6 +12,6 @@ import java.util.List;
  * @since 2024/2/21 17:17
  */
 public interface ICourseDao {
-    List<Integer> getRecommendCourse(String func, int num);
+    List<Long> getRecommendCourse(String func, int num);
     Course getCourse(int courseId);
 }
